requirements.txt
django>=3.2
djangorestframework>=3.5.2
django-mysql>=3.9
django-cors-headers>=3.5
pika>=1.1.0
mysqlclient >=2.1.1
pymysql
Flask>=1.1.2
Flask-SQLAlchemy>=2.4.4
SQLAlchemy>=1.3.20
Flask-Migrate>=2.5.3
Flask-Script>=2.0.6
Flask-Cors>=3.0.9
requests>=2.25.0
cryptography==38.0.3
django-cors-headers==3.13.0